♻️ Refactor: 훅으로 분리하여 Post페이지 코드 리팩토링
